% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestComb.R
\name{BestComb}
\alias{BestComb}
\title{BestComb}
\usage{
BestComb(AFLP, mini, maxi, who = "_R", keep = F, thresh = 80)
}
\arguments{
\item{AFLP:}{an AFLP profile loaded into RawGeno}

\item{mini:}{from where to start the analysis of the chromatogram in bp (i.e. 100)}

\item{maxi:}{to where analyse the cromatogram in bp (i.e. 400)}

\item{who:}{characters defining the replicates as define in RawGeno (i.e. "_R")}

\item{keep:}{whether to keep (T) or not (F, defaut) non replicable bins}

\item{thresh:}{threshold to evaluate reproducivitlity as implemented in RawGeno (i.e. 80, default)}
}
\description{
Search for a "Best" combination in a partitioned analysis of AFLP crhomatograms.
}
\details{
There are several ovservations that peaks which represents bigger DNA fragments
in a AFLP crhomatograms had lower intensities \insertCite{questiau1999amplified,papa2005genome,trybush2006getting,whitlock2008objective,herrmann2010selection}{CookedGeno}
Therefore, the function BestComb() allow to test partitioned ranges. This way it is possible to search a local better fit of
parameters in order to retain a maximun of loci with minimum genotyping error.
}
\examples{
#Load data into RawGeno
          BestComb(AFLP,100,400)

}
\references{
\insertAllCited{}
}
\author{
Sergio E. Sclovich
}
