% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IbinCalc.R
\name{IbinCalc}
\alias{IbinCalc}
\title{To evaluate the adequacy of the set of parameters the information content per bin (Ibin)
was used (Arrigo et al., 2009)}
\usage{
IbinCalc(matrix)
}
\description{
To evaluate the adequacy of the set of parameters the information content per bin (Ibin)
was used (Arrigo et al., 2009)
}
