% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamGeno.R
\name{ParamGeno}
\alias{ParamGeno}
\title{ParamGeno}
\usage{
ParamGeno(AFLP, case)
}
\arguments{
\item{AFLP:}{An AFLP object from Package RawGeno}

\item{case:}{A vector with the data section to be evaluated.
             i.e. if you want to evaluate parameters from 100 to 250 bp
             the vector should be:

             case = c(100,250).}
}
\description{
Exact search for "best" parameters for AFLP binary matrix composition with
package RawGeno V 2.0  \insertCite{arrigo2012automated}{CookedGeno}.
}
\details{
Recomendations from \insertCite{arrigo2012automated}{CookedGeno} has been followed with an arbitrary discretization of the ranges as follow:

* Maximun bin with (MaxBin): from 1.5 to 2.0 by 0.1 increase

* Minimun bin with (MinBin): from 1.0 to 1.5 by 0.1 increase

* Relative fluorescens units treshold (cutRFU) for the normalized intensity average:
from 50 to 200 by 25 increase.

* Reproducibility filter (TRESH): 80

It requiers RawGeno to be running and the AFLP data already loaded.
}
\examples{
#Load data into RawGeno
          ParamGeno(AFLP,c(100,500))

}
\references{
\insertAllCited{}
}
\author{
Sergio E. Sclovich
}
