% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_graf.R
\name{pcoa_graf}
\alias{pcoa_graf}
\title{pcoa_graf}
\usage{
pcoa_graf(filename, angle, x, y, z, nomb)
}
\arguments{
\item{x, y, z:}{Which eigenvector will be plot in each axis}

\item{filename:}{a pcoa object}

\item{angle:}{angle for the graph to be plot}

\item{nomb:}{Vector of names}
}
\description{
make a 3D plot based on scatterplot3d \insertCite{ligges2002scatterplot3d}{CookedGeno}
of a pcoa object from package APE  \insertCite{paradis2004ape}{CookedGeno}
}
\details{
Colors and simbols for the objects are chosen randomly
}
\examples{
pcoa_graf(pcoa_object, 120, 1, 2, 3, names(Data))
}
\references{
\insertAllCited{}
}
